Milestone - 1./requirements.txt
numpy
pandas
scikit-learn
opencv-python
pillow
scikit-image
pywavelets
joblib
tqdm
matplotlib
jupyter